move web to api dir & update vercel.json
