+++ .vscode/launch.json
@@ -12,7 +12,7 @@
       "program": "${workspaceFolder}",
       "env": {
         "PORT": "8899",
-        "Go_Proxy_BingAI_Debug": "true",
+        "Go_Proxy_BingAI_Debug": "false",
         //"Go_Proxy_BingAI_SOCKS_URL": "192.168.0.88:1070",
         // "Go_Proxy_BingAI_SOCKS_USER": "xxx",
         // "Go_Proxy_BingAI_SOCKS_PWD": "xxx",
